Missed call logs synchronization. Process AcceptedElsewhere and DeclinedElsewhere for call log synchronization Enums to String
Add an API to clone a call log.
Fix updated call log in cache: Rename main db function to match its behavior and add an out parameter to know if the call log has been updated or not.
On already stored call log, ignore different status because it is supposed that it is not targeted for the current account/device.
Improve call log synchronization tester.